Fixed incorrect terminology.
No contrast is when every thing is one color.  Perfect/Full contrast is black against white (impossible in true light because you can always have more photons coming at you though your eye may not be able to discern the difference)

"Contrast is the difference in luminance or colour that makes an object (or its representation in an image or display) distinguishable." (https://en.wikipedia.org/wiki/Contrast_(vision)).

Higher contrast is when colors are more different.  Lower contrast is when colors are more similar.